+++ src/cg/disassembler/disassembler.cpp
@@ -435,7 +435,7 @@ tuple<string, viua::internals::types::bytecode_size> disassembler::instruction(
         case BOOL:
             ptr = disassemble_ri_operand(oss, ptr);
             break;
-        case ISTORE:
+        case INTEGER:
             ptr = disassemble_ri_operand_with_rs_type(oss, ptr);
             ptr = disassemble_ri_operand(oss, ptr);