+++ src/cg/lex.cpp
@@ -762,7 +762,7 @@ namespace viua {
                             tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(),
                                                 "current");
                         }
-                    } else if (token == "istore") {
+                    } else if (token == "integer") {
                         tokens.push_back(token);                 // mnemonic
                         tokens.push_back(input_tokens.at(++i));  // target register
                         if (not is_register_set_name(input_tokens.at(i + 1))) {
@@ -1435,7 +1435,7 @@ namespace viua {
                             tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(),
                                                 "current");
                         }
-                    } else if (token == "istore") {
+                    } else if (token == "integer") {
                         tokens.push_back(input_tokens.at(++i));  // target register
                         if (not is_register_set_name(input_tokens.at(i + 1))) {
                             tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(),
@@ -2374,7 +2374,7 @@ namespace viua {
                         tokens.back().original("default");
                         continue;
                     }
-                    if (match(input_tokens, i, {"istore", "", "default"})) {
+                    if (match(input_tokens, i, {"integer", "", "default"})) {
                         tokens.push_back(input_tokens.at(++i));  // push target register token
                         ++i;
                         tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(), "0");