+++ src/front/asm/assemble_instruction.cpp
@@ -94,7 +94,7 @@ static tuple<viua::internals::types::bytecode_size, enum JUMPTYPE> resolvejump(
 
 static string resolveregister(Token token, const bool allow_bare_integers = false) {
     /*  This function is used to register numbers when a register is accessed, e.g.
-     *  in `istore` instruction or in `branch` in condition operand.
+     *  in `integer` instruction or in `branch` in condition operand.
      *
      *  This function MUST return string as teh result is further passed to assembler::operands::getint()
      * function which *expects* string.
@@ -223,11 +223,11 @@ viua::internals::types::bytecode_size assemble_instruction(
 
         program.opizero(assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)),
                                                                  resolve_rs_type(tokens.at(target + 1))));
-    } else if (tokens.at(i) == "istore") {
+    } else if (tokens.at(i) == "integer") {
         TokenIndex target = i + 1;
         TokenIndex source = target + 2;
 
-        program.opistore(assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)),
+        program.opinteger(assembler::operands::getint_with_rs_type(resolveregister(tokens.at(target)),
                                                                   resolve_rs_type(tokens.at(target + 1))),
                          assembler::operands::getint(resolveregister(tokens.at(source), true), true));
     } else if (tokens.at(i) == "iinc") {